+++ src/components/NewThoughtForm.jsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 
-export const ThoughtsForm = ({ setThoughts, fetchData, apiUrl }) => {
+export const NewThoughtForm = ({ setThoughts, fetchData, apiUrl }) => {
   const [message, setMessage] = useState("");
 
   const handleChange = (event) => {
@@ -23,14 +23,18 @@ export const ThoughtsForm = ({ setThoughts, fetchData, apiUrl }) => {
   };
 
   return (
-    <form className="input-box" onSubmit={handleSubmit}>
+    <form className="box input-box" onSubmit={handleSubmit}>
       <p>What's making you happy right now?</p>
       <textarea
         type="text"
         placeholder="Write a happy thought"
+        id="new-thought"
         value={message}
         onChange={handleChange}
       />
+      <label htmlFor="new-thought" className="visually-hidden">
+        Write a happy thought
+      </label>
       <button className="submit-button" type="submit">
         ❤️ Send Happy Thought ❤️
       </button>
