+++ src/index.css
@@ -9,23 +9,25 @@
   --button-grey: #eae7e7;
 }
 
+code {
+  font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
+    monospace;
+}
+
 * {
   box-sizing: border-box;
 }
 
-code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
-    monospace;
+*:focus {
+  outline: 2px purple solid;
 }
 
 body {
   font-family: "Montserrat";
   font-weight: 800;
 }
-
-span {
-  color: gray;
-  font-size: 0.7rem;
+p {
+  margin: 10px 0;
 }
 
 textarea {
@@ -39,19 +41,49 @@ textarea {
   display: flex;
   flex-direction: column;
   gap: 2rem;
+  align-items: center;
 }
 
-.input-box {
+.thoughts-container {
+  display: flex;
+  flex-direction: column;
+  gap: 2rem;
+}
+
+.box {
   display: flex;
   flex-direction: column;
   border: solid;
-  width: 400px;
+  min-width: 310px;
+  max-width: 345px;
   box-shadow: black 5px 5px;
+}
+
+.input-box {
   padding: 10px 20px;
   background-color: #f2f0f0;
   gap: 5px;
+  margin-top: 10px;
 }
 
+.thought-box {
+  padding: 20px;
+  gap: 10px;
+}
+
+.likes-timestamp {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  color: gray;
+  font-size: 0.7rem;
+}
+
+.timestamp {
+  color: #cecdd3;
+}
+
+/* ... Buttons ... */
 button {
   width: fit-content;
   cursor: pointer;
@@ -67,11 +99,18 @@ button {
   font-weight: 600;
 }
 
+.submit-button:hover {
+  box-shadow: inset 0 0 0 2px red;
+}
+
 .likes-button {
   padding: 10px;
   border-radius: 50%;
-
   background-color: var(--button-grey);
+}
+
+.likes-button:hover {
+  background-color: var(--pink);
 }
 
 .likes-button:disabled {
@@ -81,28 +120,22 @@ button {
   cursor: default;
 }
 
-.message-container {
-  display: flex;
-  flex-direction: column;
-  gap: 2rem;
+/* ... Visually hidden ... */
+.visually-hidden {
+  border: 0;
+  clip: rect(0 0 0 0);
+  height: 1px;
+  margin: -1px;
+  overflow: hidden;
+  padding: 0;
+  position: absolute;
+  width: 1px;
 }
 
-.message-box {
-  display: flex;
-  flex-direction: column;
-  border: solid;
-  width: 400px;
-  padding: 20px;
-  box-shadow: black 5px 5px;
-  gap: 10px;
-}
+/***** Tablet & Desktop *****/
 
-.likes-time {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-button:hover {
-  background-color: salmon;
+@media (min-width: 768px) {
+  .box {
+    min-width: 500px;
+  }
 }
